fixed linting problems
