+++ .eslintrc.js
@@ -20,5 +20,7 @@ module.exports = {
     'prettier',
   ],
   rules: {
+    "react/prop-types": "off",
+    "react/jsx-filename-extension": [1, { "extensions": [".js", ".jsx"] }]
   },
 };
