+++ src/App.js
@@ -1,3 +1,4 @@
+import React from "react";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import LoginView from "./pages/loginView";
 import HomeView from "./pages/homeView";
@@ -10,7 +11,7 @@ function App() {
       <div className="background">
         <div className="main-content">
           <Switch>
-            {/* Login/Signup page - no navbar*/}
+            {/* Login/Signup page - no navbar */}
             <Route exact path="/">
               <LoginView />
             </Route>
@@ -20,7 +21,7 @@ function App() {
               <Navbar />
               <HomeView />
             </Route>
-            
+
             {/* Fallback - if none of the above routes are hit */}
             <Route>
               <PageNotFoundView />
