+++ src/components/Navbar.jsx
@@ -1,10 +1,10 @@
 import React, { useState } from "react";
+import { useLocation, Link } from "react-router-dom";
 import BoardIcon from "../assets/icons/BoardIcon";
 import DashboardIcon from "../assets/icons/DashboardIcon";
 import HomeIcon from "../assets/icons/HomeIcon";
 import SettingsIcon from "../assets/icons/SettingsIcon";
 import LogoutIcon from "../assets/icons/LogoutIcon";
-import { useLocation, Link } from "react-router-dom";
 import DefaultAvatar from "../assets/images/default-avatar.png";
 
 function Navbar() {
@@ -13,13 +13,13 @@ function Navbar() {
   return (
     <div className="navbar">
       <div>
-        <Link to={"/home"} className="button-link">
+        <Link to="/home" className="button-link">
           <div className="navbar__icon__group">
             {location.pathname === "/home" && <div className="selected-box" />}
             <HomeIcon />
           </div>
         </Link>
-        <Link to={"/dashboard"} className="button-link">
+        <Link to="/dashboard" className="button-link">
           <div className="navbar__icon__group">
             {location.pathname === "/dashboard" && (
               <div className="selected-box" />
@@ -27,7 +27,7 @@ function Navbar() {
             <DashboardIcon />
           </div>
         </Link>
-        <Link to={"/board"} className="button-link">
+        <Link to="/board" className="button-link">
           <div className="navbar__icon__group">
             {location.pathname === "/board" && <div className="selected-box" />}
             <BoardIcon />
@@ -40,7 +40,7 @@ function Navbar() {
         onMouseLeave={() => setHovering(false)}
       >
         <div
-          className={"navbar__account" + (hovering ? "--hover" : "")}
+          className={`navbar__account${  hovering ? "--hover" : ""}`}
         >
           <SettingsIcon />
           <LogoutIcon />
