+++ src/components/global/Button.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { getIcon } from "./componentFunctions";
 import { Link } from "react-router-dom";
+import getIcon from "./componentFunctions";
 
 function Button(props) {
   const {path, text, icon, height, fontSize} = props;
