+++ src/components/global/Task/Subtask.jsx
@@ -10,10 +10,11 @@ function Subtask(props) {
   return (
     <div className="subtask-container">
       <div
-        className={"subtask-checkbox" + (isChecked ? "--checked" : "")}
+        className={`subtask-checkbox${  isChecked ? "--checked" : ""}`}
         onClick={handleCheckBoxClick}
+        onKeyDown={handleCheckBoxClick}
       />
-      <span className={"subtask-title" + (isChecked ? "--checked" : "")}>
+      <span className={`subtask-title${  isChecked ? "--checked" : ""}`}>
         {name}
       </span>
     </div>
