+++ src/components/global/TextInput.jsx
@@ -1,9 +1,12 @@
+import React from "react";
+
 function TextInput(props) {
   const { label, textValue, placeholderValue, onChangeHandler } = props;
   
   return (
     <div className="text-input">
       {label && 
+        // eslint-disable-next-line jsx-a11y/label-has-associated-control
         <label className="text-input__label">
           {label}
         </label>
