+++ src/pages/homeView.js
@@ -4,7 +4,7 @@ import Button from "../components/global/Button";
 function homeView() {
   return (
     <div>
-      <Button icon={"rightArrow"} text={"Dashboard"} path={"/home"} />
+      <Button icon="rightArrow" text="Dashboard" path="/home" />
     </div>
   );
 }
