+++ src/pages/loginView.js
@@ -1,4 +1,6 @@
-export default function LoginView( ) {
+import React from "react";
+
+export default function LoginView() {
     return (
         <div>
             <h2>Login</h2>
