+++ src/pages/pageNotFoundView.js
@@ -1,4 +1,6 @@
-export default function PageNotFoundView( ) {
+import React from "react";
+
+export default function PageNotFoundView() {
     return (
         <div>
             <h2>ERROR</h2>
